+++ packages/clients/dish/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: If a user is geolocated outside of the map's extent, the client will inform the user of why geolocation did not take effect via a textbox.
 - Fix: Extend typing for search result function according to type package update.
+- Fix: Import types `AddressSearchState` and `AddressSearchGetters` from correct position.
 - Chore: Change value of `pins.movable` configuration to `'drag'` as using a boolean has been deprecated in a future release.
 
 ## 1.1.1