+++ packages/plugins/AddressSearch/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Feature: Add new optional configuration parameter `afterResultComponent` that allows to display a custom component for each search result.
 - Feature: There is now a new optional configuration parameter `searchMethods.queryParameters.likeFilterAttributes` for the WfsSearch that allows to add custom key/value pairs which are mapped to attributes of the like filter operator.
 - Fix: Adjust documentation and types to properly describe optionality of configuration parameters.
+- Fix: `AddressSearchState` and `AddressSearchGetters` types have been additionally exposed as root export since using packages frequently rely on them.
 
 ## 1.2.1